feat: implement dashboard header version toggle based on feature flag
